Only allow 'Default' for Discord targets
Use the hard-coded string "Default" for all discord targets

Also deduplicate isTargetsExist check
