+++ packages/notifi-react-card/lib/hooks/useIsTargetsExist.ts
@@ -1,0 +1,19 @@
+import { useMemo } from 'react';
+
+import { useNotifiSubscriptionContext } from '../context';
+
+export const useIsTargetsExist = () => {
+  const { useDiscord, email, phoneNumber, telegramId, discordTargetData } =
+    useNotifiSubscriptionContext();
+
+  const isTargetsExist = useMemo(() => {
+    return (
+      !!email ||
+      !!phoneNumber ||
+      !!telegramId ||
+      (useDiscord && !!discordTargetData?.id)
+    );
+  }, [email, phoneNumber, telegramId, discordTargetData, useDiscord]);
+
+  return isTargetsExist;
+};
